Add FlywayDB migrations

+++ newm-server/src/main/kotlin/io/newm/server/auth/twofactor/database/TwoFactorAuthEntity.kt
@@ -7,13 +7,13 @@ import org.jetbrains.exposed.sql.deleteWhere
 import org.jetbrains.exposed.sql.lowerCase
 import java.time.LocalDateTime
 
-class TwoFactorAuthEntity(id: EntityID<Int>) : Entity<Int>(id) {
+class TwoFactorAuthEntity(id: EntityID<Long>) : Entity<Long>(id) {
 
     var email by TwoFactorAuthTable.email
     var codeHash by TwoFactorAuthTable.codeHash
     var expiresAt by TwoFactorAuthTable.expiresAt
 
-    companion object : EntityClass<Int, TwoFactorAuthEntity>(TwoFactorAuthTable) {
+    companion object : EntityClass<Long, TwoFactorAuthEntity>(TwoFactorAuthTable) {
 
         fun getByEmail(email: String): TwoFactorAuthEntity? = find {
             TwoFactorAuthTable.email.lowerCase() eq email.lowercase()
